refactor heads

+++ ppocr/modeling/heads/table_master_head.py
@@ -1,16 +1,3 @@
-# Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
 """
 This code is refer from:
 https://github.com/JiaquanYe/TableMASTER-mmocr/blob/master/mmocr/models/textrecog/decoders/master_decoder.py
@@ -18,6 +5,7 @@ https://github.com/JiaquanYe/TableMASTER-mmocr/blob/master/mmocr/models/textreco
 
 import copy
 import math
+
 import torch
 from torch import nn
 from torch.nn import functional as F
@@ -213,7 +201,7 @@ class SubLayerConnection(nn.Module):
 
 def masked_fill(x, mask, value):
     mask = mask.astype(x.dtype)
-    return x * torch.logical_not(mask).astype(x.dtype) + mask * value
+    return x * torch.logical_not(mask).type(x.dtype) + mask * value
 
 
 def self_attention(query, key, value, mask=None, dropout=None):
@@ -259,7 +247,7 @@ class PositionalEncoding(nn.Module):
 
         # Compute the positional encodings once in log space.
         pe = torch.zeros([max_len, d_model])
-        position = torch.arange(0, max_len).unsqueeze(1).astype("float32")
+        position = torch.arange(0, max_len).unsqueeze(1).astype(torch.float32)
         div_term = torch.exp(torch.arange(0, d_model, 2) * -math.log(10000.0) / d_model)
         pe[:, 0::2] = torch.sin(position * div_term)
         pe[:, 1::2] = torch.cos(position * div_term)
